\name{simple.band1}

\docType{data}
\alias{simple.band1}

\title{
  Simulated simple data of a doubly-lensed quasar observed in band 1
}

\description{
  A 50 by 5 matrix containing a data set of a doubly-lensed quasar observed in band 1, which is modified from realistic data. 
}

\usage{data(simple.band1)}

\format{
  A simple data set of a doubly-lensed quasar:
  \describe{
    \item{\code{time}}{observation times}
    \item{\code{x}}{magnitudes of light A observed in band 1}
    \item{\code{se.x}}{measurement errors of x}
    \item{\code{y}}{magnitudes of light B observed in band 1}
    \item{\code{se.y}}{measurement errors of y}
  }
}

\examples{
  data(simple.band1)
}

\keyword{datasets}
