\name{simple.band2}

\docType{data}
\alias{simple.band2}

\title{
  Simulated simple data of a doubly-lensed quasar observed in band 2
}

\description{
  A 30 by 5 matrix containing a data set of a doubly-lensed quasar observed in band 2, which is modified from realistic data. 
}

\usage{data(simple.band2)}

\format{
  A simple data set of a doubly-lensed quasar:
  \describe{
    \item{\code{time}}{observation times}
    \item{\code{x}}{magnitudes of light A observed in band 2}
    \item{\code{se.x}}{measurement errors of x}
    \item{\code{y}}{magnitudes of light B observed in band 2}
    \item{\code{se.y}}{measurement errors of y}
  }
}

\examples{
  data(simple.band2)
}

\keyword{datasets}
