\name{timedelay-internal}
\alias{postTheta}
\alias{logpostDelta}
\alias{postX}
\alias{Log.post.theta}
\alias{postTheta.multi}
\alias{postBeta}
\alias{Log.post.delta}
\alias{postX.multi}
\alias{orderData}

\title{Internal \code{bayesian} functions}
\description{Internal \code{bayesian} functions}
\details{These are not to be called by the user.}
\keyword{internal}

